# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation My API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/posts/sort-by-newest": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get Posts Data sorted by newest created_date",
                "operationId": "7977f1adf19cc98922b5a7a07efdb104",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Get Posts Data"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": 1
                                                    },
                                                    "user_id": {
                                                        "type": "number",
                                                        "example": 1
                                                    },
                                                    "body": {
                                                        "type": "string",
                                                        "example": "testing komen pakai github"
                                                    },
                                                    "likes": {
                                                        "type": "number",
                                                        "example": 148
                                                    },
                                                    "total_comments": {
                                                        "type": "number",
                                                        "example": 0
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2024-02-15T14:29:30.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2024-02-16T10:23:15.000000Z"
                                                    },
                                                    "comments": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "number",
                                                                    "example": 1
                                                                },
                                                                "user_id": {
                                                                    "type": "number",
                                                                    "example": 1
                                                                },
                                                                "post_id": {
                                                                    "type": "number",
                                                                    "example": 1
                                                                },
                                                                "body": {
                                                                    "type": "string",
                                                                    "example": "testing comment"
                                                                },
                                                                "created_at": {
                                                                    "type": "string",
                                                                    "example": null
                                                                },
                                                                "updated_at": {
                                                                    "type": "string",
                                                                    "example": null
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "failed"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}